server-old/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==3.0.3
Flask-Cors==5.0.0
numpy==2.0.2
pillow==10.4.0
psutil==6.0.0
scikit-image==0.24.0
scikit-learn==1.5.2
trimesh==4.4.9
opencv-python==4.10.0.84

# Leave here
gunicorn==23.0.0
fast_simplification==0.1.9